#20 making the code clean, version updates
