+++ app/build.gradle
@@ -32,7 +32,7 @@ android {
     }
     kotlinOptions {
         jvmTarget = "1.8"
-        useIR = true
+        freeCompilerArgs = ['-Xjvm-default=enable']
     }
     buildFeatures {
         compose true
@@ -61,7 +61,7 @@ dependencies {
     implementation "androidx.lifecycle:lifecycle-runtime-ktx:$rootProject.ext.versions.lifecycle"
     implementation "androidx.activity:activity-compose:$rootProject.ext.versions.activityCompose"
     implementation "androidx.lifecycle:lifecycle-viewmodel-compose:$rootProject.ext.versions.lifecycle"
-    implementation "androidx.navigation:navigation-compose:2.4.2"
+    implementation "androidx.navigation:navigation-compose:2.5.0"
     implementation "com.google.accompanist:accompanist-pager:0.19.0"
     implementation "com.google.accompanist:accompanist-pager-indicators:0.19.0"
     implementation "androidx.compose.runtime:runtime-livedata:$rootProject.ext.versions.compose"
