+++ app/src/main/java/ru/frozenpriest/pokebase/di/AppComponentHolder.kt
@@ -1,6 +1,6 @@
 package ru.frozenpriest.pokebase.di
 
-import ru.frozenpriest.pokebase.data.local.DataStoreRepository
+import ru.frozenpriest.pokebase.domain.DataStoreRepository
 import ru.frozenpriest.pokebase.domain.login.LoginRegisterUseCase
 import ru.frozenpriest.pokebase.domain.pokemon.AddPokemonUseCase
 import ru.frozenpriest.pokebase.domain.pokemon.GetDamageUseCase
