+++ app/src/main/java/ru/frozenpriest/pokebase/presentation/screens/login/LoginRegisterViewModel.kt
@@ -7,7 +7,7 @@ import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.receiveAsFlow
 import kotlinx.coroutines.launch
-import ru.frozenpriest.pokebase.data.local.DataStoreRepository
+import ru.frozenpriest.pokebase.domain.DataStoreRepository
 import ru.frozenpriest.pokebase.domain.login.LoginRegisterUseCase
 import timber.log.Timber
 import javax.inject.Inject
