--- data/src/main/java/ru/frozenpriest/pokebase/data/local/LocalRepository.kt
@@ -1,3 +1,0 @@
-package ru.frozenpriest.pokebase.data.local
-
-interface LocalRepository
