+++ data/src/main/java/ru/frozenpriest/pokebase/data/remote/model/TokenResponse.kt
@@ -2,9 +2,12 @@ package ru.frozenpriest.pokebase.data.remote.model
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
+import ru.frozenpriest.pokebase.domain.model.Token
 
 @Serializable
 data class TokenResponse(
     @SerialName("JWT")
     val token: String
 )
+
+fun TokenResponse.toToken(): Token = Token(token)
