+++ domain/build.gradle
@@ -32,13 +32,9 @@ android {
 
 dependencies {
     implementation project(path: ':module_injector')
+    implementation 'androidx.annotation:annotation:1.4.0'
 
 
-    implementation 'androidx.core:core-ktx:1.7.0'
-    implementation 'androidx.appcompat:appcompat:1.4.1'
-    implementation 'com.google.android.material:material:1.6.0'
-    implementation 'androidx.compose.ui:ui-graphics:1.1.1'
-    implementation project(path: ':data')
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
