+++ domain/src/main/java/ru/frozenpriest/pokebase/domain/DataStoreRepository.kt
@@ -1,0 +1,6 @@
+package ru.frozenpriest.pokebase.domain
+
+interface DataStoreRepository {
+    suspend fun getBearerToken(): String
+    suspend fun setBearerToken(token: String)
+}
