+++ domain/src/main/java/ru/frozenpriest/pokebase/domain/di/DomainModule.kt
@@ -2,7 +2,7 @@ package ru.frozenpriest.pokebase.domain.di
 
 import dagger.Module
 import dagger.Provides
-import ru.frozenpriest.pokebase.data.remote.RemoteRepository
+import ru.frozenpriest.pokebase.domain.RemoteRepository
 import ru.frozenpriest.pokebase.domain.login.LoginRegisterUseCase
 import ru.frozenpriest.pokebase.domain.login.LoginRegisterUseCaseImpl
 import ru.frozenpriest.pokebase.domain.pokemon.AddPokemonUseCase
