+++ domain/src/main/java/ru/frozenpriest/pokebase/domain/model/Move.kt
@@ -1,7 +1,5 @@
 package ru.frozenpriest.pokebase.domain.model
 
-import ru.frozenpriest.pokebase.data.remote.model.MoveResponse
-
 data class Move(
     val id: String,
     val name: String,
@@ -11,15 +9,3 @@ data class Move(
     val accuracy: Float?,
     val pp: Int,
 )
-
-fun MoveResponse.toMove(): Move {
-    return Move(
-        id = moveId.toString(),
-        name = name,
-        type = Type.valueOf(type),
-        category = Category.valueOf(category),
-        power = power,
-        accuracy = accuracy?.toFloat(),
-        pp = pp
-    )
-}
