+++ domain/src/main/java/ru/frozenpriest/pokebase/domain/model/PokemonShort.kt
@@ -1,7 +1,5 @@
 package ru.frozenpriest.pokebase.domain.model
 
-import ru.frozenpriest.pokebase.data.remote.model.PokemonShortResponse
-
 data class PokemonShort(
     val id: String,
     val name: String,
@@ -9,13 +7,3 @@ data class PokemonShort(
     val imageUrl: String,
     val types: List<Type>,
 )
-
-fun PokemonShortResponse.toPokemonShort(): PokemonShort {
-    return PokemonShort(
-        id = pokemonId,
-        name = nickname,
-        speciesName = species.name,
-        imageUrl = species.image,
-        types = listOfNotNull(species.primaryType, species.secondaryType).map { Type.valueOf(it) }
-    )
-}
