+++ domain/src/main/java/ru/frozenpriest/pokebase/domain/model/Token.kt
@@ -1,0 +1,5 @@
+package ru.frozenpriest.pokebase.domain.model
+
+data class Token(
+    val token: String
+)
