+++ domain/src/main/java/ru/frozenpriest/pokebase/domain/pokemon/GetMovesUseCase.kt
@@ -1,9 +1,8 @@
 package ru.frozenpriest.pokebase.domain.pokemon
 
-import ru.frozenpriest.pokebase.data.remote.RemoteRepository
+import ru.frozenpriest.pokebase.domain.RemoteRepository
 import ru.frozenpriest.pokebase.domain.model.Move
 import ru.frozenpriest.pokebase.domain.model.Species
-import ru.frozenpriest.pokebase.domain.model.toMove
 import javax.inject.Inject
 
 interface GetMovesUseCase {
@@ -16,9 +15,7 @@ class GetMovesUseCaseImpl @Inject constructor(
     private val pokeBaseRepository: RemoteRepository
 ) : GetMovesUseCase {
     override suspend fun getMoves(pokemonSpecies: Species): Result<List<Move>> {
-        return pokeBaseRepository.getMoves(pokemonSpecies.id).map { list ->
-            list.map { movesResponse -> movesResponse.toMove() }
-        }
+        return pokeBaseRepository.getMoves(pokemonSpecies.id)
     }
 
     override suspend fun addMove(pokemonId: String, moveId: String) {
