+++ domain/src/main/java/ru/frozenpriest/pokebase/domain/pokemon/GetOwnedPokemonShortUseCase.kt
@@ -1,8 +1,7 @@
 package ru.frozenpriest.pokebase.domain.pokemon
 
-import ru.frozenpriest.pokebase.data.remote.RemoteRepository
+import ru.frozenpriest.pokebase.domain.RemoteRepository
 import ru.frozenpriest.pokebase.domain.model.PokemonShort
-import ru.frozenpriest.pokebase.domain.model.toPokemonShort
 import javax.inject.Inject
 
 interface GetOwnedPokemonShortUseCase {
@@ -13,8 +12,6 @@ class GetOwnedPokemonShortUseCaseImpl @Inject constructor(
     private val pokeBaseRepository: RemoteRepository
 ) : GetOwnedPokemonShortUseCase {
     override suspend fun getPokemon(): Result<List<PokemonShort>> {
-        return pokeBaseRepository.getOwnedPokemon().map { list ->
-            list.map { pokemonResponse -> pokemonResponse.toPokemonShort() }
-        }
+        return pokeBaseRepository.getOwnedPokemon()
     }
 }
