+++ domain/src/main/java/ru/frozenpriest/pokebase/domain/pokemon/GetPokemonDetailsUseCase.kt
@@ -1,8 +1,7 @@
 package ru.frozenpriest.pokebase.domain.pokemon
 
-import ru.frozenpriest.pokebase.data.remote.RemoteRepository
+import ru.frozenpriest.pokebase.domain.RemoteRepository
 import ru.frozenpriest.pokebase.domain.model.Pokemon
-import ru.frozenpriest.pokebase.domain.model.toPokemon
 import javax.inject.Inject
 
 interface GetPokemonDetailsUseCase {
@@ -14,6 +13,5 @@ class GetPokemonDetailsUseCaseImpl @Inject constructor(
 ) : GetPokemonDetailsUseCase {
     override suspend fun getPokemonDetails(pokemonId: String): Result<Pokemon> {
         return pokeBaseRepository.getPokemonDetails(pokemonId)
-            .map { pokemonDetailsResponse -> pokemonDetailsResponse.toPokemon() }
     }
 }
