+++ domain/src/main/java/ru/frozenpriest/pokebase/domain/pokemon/GetSpeciesUseCase.kt
@@ -1,8 +1,7 @@
 package ru.frozenpriest.pokebase.domain.pokemon
 
-import ru.frozenpriest.pokebase.data.remote.RemoteRepository
+import ru.frozenpriest.pokebase.domain.RemoteRepository
 import ru.frozenpriest.pokebase.domain.model.SpeciesShort
-import ru.frozenpriest.pokebase.domain.model.toShort
 import javax.inject.Inject
 
 interface GetSpeciesUseCase {
@@ -13,8 +12,6 @@ class GetSpeciesUseCaseImpl @Inject constructor(
     private val pokeBaseRepository: RemoteRepository
 ) : GetSpeciesUseCase {
     override suspend fun getSpecies(): Result<List<SpeciesShort>> {
-        return pokeBaseRepository.getSpecies().map { list ->
-            list.map { speciesResponse -> speciesResponse.toShort() }
-        }
+        return pokeBaseRepository.getSpecies()
     }
 }
